Import of RKOBES/PPM-Make-0.88 from CPAN.
gitpan-cpan-distribution: PPM-Make
gitpan-cpan-version:      0.88
gitpan-cpan-path:         RKOBES/PPM-Make-0.88.tar.gz
gitpan-cpan-author:       RKOBES
gitpan-cpan-maturity:     released

+++ t/bundle.t
@@ -56,5 +56,5 @@ foreach my $arch (keys %exts) {
   my $zipdist = File::Spec->catfile($cwd, 'Bundle-AppConfig.zip');
   ok(-f $zipdist);
   unlink ($zipdist);
-  rmtree($build_dir, 1, 1);
+  rmtree($build_dir, 1, 1) if (defined $build_dir and -d $build_dir);
 }
